Version of software: V0.01.01
Commit details:
Programmed the PWM generation using interrupt and added few changes to other files too.

Observations:
PWM is not generated from the enabled GPIO pins. And the duty cycle for disabled pins are not according to the set value.

TODOs:
To solve the duty cycle issue and the frequency issue.